Safari-611.1.10/LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/css/cssom-view/Screen-pixelDepth-Screen-colorDepth001-expected.txt

This case tests the Screen pixelDepth and colorDepth

The test passes if the value is 24


FAIL testColorDepth assert_equals: Expected value for colorDepth is 24 expected 24 but got 32
FAIL testPixelDepth assert_equals: Expected value for pixelDepth is 24 expected 24 but got 32
